requirements.txt
requests
pandas
plotly
python-dotenv
beautifulsoup4
lxml
